Lines of Action
An abstract strategy board game in Java, modeled after the "Lines of Action" board game by Claude Soucie. The objective is to connect all of one's pieces into a single group. The program also implements an opposing AI with a heuristic and minimax algorithm designed to beat the human player within a set number of moves.